requirements.txt
python-dotenv>=1.0
pytest>=9.0
